Cambios en RoomDAO

+++ src/main/java/repository/dao/RoomDAO.java
@@ -4,6 +4,7 @@ import model.Room;
 import exception.PersistenceException;
 import repository.database.DatabaseConfig;
 import repository.mapper.RoomMapper;
+import repository.mapper.GeneralMapper;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,7 +12,7 @@ import java.util.Optional;
 
     public class RoomDAO implements GenericDAO<Room, Long> {
 
-        private final RoomMapper mapper = RoomMapper.getInstance();
+        private final GeneralMapper<Room> mapper = RoomMapper.getInstance();
 
         @Override
         public Room save(Room room) {
@@ -28,46 +29,57 @@ import java.util.Optional;
                 }
 
                 setGeneratedId(room, stmt);
-                return room;
+
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al guardar la sala: " + room.getName() + ".");
             }
+            return room;
         }
 
         @Override
         public Optional<Room> findById(Long id) {
-            String sql = "SELECT id, name, difficulty_level, price, escape_room_id FROM room WHERE id = ?";
+            if (id == null || id <= 0){
+                throw new IllegalArgumentException("El id debe ser un numero positivo.");
+            }
 
+            String sql = "SELECT id, name, difficulty_level, price, escape_room_id FROM room WHERE id = ?";
+            Optional<Room> room = Optional.empty();
             try (Connection conn = DatabaseConfig.getConnection();
                  PreparedStatement stmt = conn.prepareStatement(sql)) {
 
                 stmt.setLong(1, id);
-                return executeQueryAndMapToRoom(stmt);
+                room =  executeQueryAndMapToRoom(stmt);
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al buscar la sala con ID = " + id + ".");
             }
+            return  room;
         }
 
         @Override
         public Optional<Room> findByName(String name) {
-            String sql = "SELECT id, name, difficulty_level, price FROM room WHERE name = ?";
+            if (name == null || name.trim().isEmpty()) {
+                throw new IllegalArgumentException("El nombre no puede estar vacío.");
+            }
+            String sql = "SELECT id, name, difficulty_level, price , escape_room-id FROM room WHERE name = ?";
+            Optional<Room> room = Optional.empty();
 
             try (Connection conn = DatabaseConfig.getConnection();
                  PreparedStatement stmt = conn.prepareStatement(sql)) {
 
                 stmt.setString(1, name);
-                return executeQueryAndMapToRoom(stmt);
+                room = executeQueryAndMapToRoom(stmt);
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al buscar sala por nombre: " + name + ".");
             }
+            return room;
         }
 
         @Override
         public List<Room> findAll() {
-            String sql = "SELECT id, name, difficulty_level, price FROM room";
+            String sql = "SELECT id, name, difficulty_level, price,escape_room_id FROM room";
             List<Room> rooms = new ArrayList<>();
 
             try (Connection conn = DatabaseConfig.getConnection();
@@ -75,30 +87,34 @@ import java.util.Optional;
                  ResultSet rs = stmt.executeQuery()) {
 
                 while (rs.next()) {
-                    Room room = mapper.fromResultSet(rs);
-                    rooms.add(room);
+                    rooms.add(mapper.fromResultSet(rs));
                 }
-                return rooms;
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al buscar todas las salas.");
             }
+            return rooms;
         }
 
         @Override
         public boolean delete(Long id) {
+            if (id == null || id <= 0) {
+                throw new IllegalArgumentException("El ID debe ser un número positivo.");
+            }
             String sql = "DELETE FROM room WHERE id = ?";
+            int affectedRows = 0;
 
             try (Connection conn = DatabaseConfig.getConnection();
                  PreparedStatement stmt = conn.prepareStatement(sql)) {
 
                 stmt.setLong(1, id);
-                int affectedRows = stmt.executeUpdate();
-                return affectedRows > 0;
+                affectedRows = stmt.executeUpdate();
+
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al borrar la sala.");
             }
+            return affectedRows > 0;
         }
 
         private void setGeneratedId(Room room, PreparedStatement stmt) throws SQLException {
@@ -121,6 +137,9 @@ import java.util.Optional;
         }
 
         public List<Room> findByEscapeRoomId(Long escapeRoomId) {
+            if (escapeRoomId == null || escapeRoomId <= 0) {
+                throw new IllegalArgumentException("El escapeRoomId debe ser un número positivo.");
+            }
             String sql = "SELECT id, name, difficulty_level, price, escape_room_id FROM room WHERE escape_room_id = ?";
             List<Room> rooms = new ArrayList<>();
 
@@ -133,14 +152,18 @@ import java.util.Optional;
                     Room room = mapper.fromResultSet(rs);
                     rooms.add(room);
                 }
-                return rooms;
 
             } catch (SQLException e) {
                 throw new PersistenceException("Error al buscar salas por escape room ID: " + escapeRoomId);
             }
+            return  rooms;
         }
             public boolean updateEscapeRoomRelation (Long roomId, Long escapeRoomId){
+                if (roomId == null || roomId <= 0) {
+                    throw new IllegalArgumentException("El roomId debe ser un número positivo.");
+                }
                 String sql = "UPDATE room SET escape_room_id = ? WHERE id = ?";
+                int affectedRows = 0;
 
                 try (Connection conn = DatabaseConfig.getConnection();
                      PreparedStatement stmt = conn.prepareStatement(sql)) {
@@ -152,12 +175,12 @@ import java.util.Optional;
                     }
                     stmt.setLong(2, roomId);
 
-                    int affectedRows = stmt.executeUpdate();
-                    return affectedRows > 0;
+                    affectedRows = stmt.executeUpdate();
 
                 } catch (SQLException e) {
                     throw new PersistenceException("Error al actualizar relación con escape room para sala ID: " + roomId);
                 }
+                return affectedRows > 0;
             }
         }
